Switched away from using matrices for the rendering, it's just doing simple offset calculations now.  This allowed me to get rid of the glm dependency.  Next dependency to get rid of is probably GLEW, but no rush on that.
Added a performance counter to the game that measures clock cycles for different systems. Tried using real clock time but the SDL clock has too low of a resolution for that. The largest performance hit is the console, as expected since it is drawing every character individually. Next is the renderer because culling isn't being done on the sprites. Collision is lower than expected, but still needs an optimization pass so it isn't checking every collider multiple times.