Cleaned up more of the Debug Console. It looks like the Renderer->Flush is taking up the time required for vsync to work, but the frame rate unlocks if swap buffers is never called. Need to understand this better.
Next up for optimization is the collision system, which takes a lot of time when many things are colliding, but that is because the whole scene is being checked. Need to implement a Broad Phase of collision detection so that the narrow phase can only check against a few static colliders.