New tilemap system
Old tilemap system and physics system that was connected to it has been
removed.

Camera entity can now follow an entity. This is backwards. There should
be a camera in the scene, and then an entity that wants to control it
can update the scene's camera. Camera as an entity is janky.