More sane entity ID system
Entities now use an index and a generation as an EntityID. EntityIDs can
be stored and used to get a pointer to an entity that will remain valid
throughout the frame, but not beyond. Get a fresh pointer every frame!
It's fast because the EntityID.id is an index into the entities array.